move MessageDigest to crypto_tools module
