+++ src/crypto_tools/mod.rs
@@ -1,6 +1,7 @@
 pub mod constants;
 pub mod hash;
 pub mod k256_serde;
+pub mod message_digest;
 pub mod mta;
 pub mod paillier;
 pub mod rng;
